fix: 阿里小程序等<svg> 需要encodeUri操作， 提取部分公共代码
