+++ src/libs/generateAlipayComponent.ts
@@ -6,13 +6,13 @@ import colors from 'colors';
 import { XmlData } from './fetchXml';
 import { Config } from './getConfig';
 import { getTemplate } from './getTemplate';
+import { generateCase } from "./utils"
 import {
   replaceIsRpx,
   replaceNames,
   replaceSize,
 } from './replace';
 
-const ATTRIBUTE_FILL_MAP = ['path'];
 
 export const generateAlipayComponent = (data: XmlData, config: Config) => {
   const svgTemplates: string[] = [];
@@ -59,53 +59,3 @@ export const generateAlipayComponent = (data: XmlData, config: Config) => {
   console.log(`\n${colors.green('√')} All icons have been putted into dir: ${colors.green(config.save_dir)}\n`);
 };
 
-const generateCase = (data: XmlData['svg']['symbol'][number]) => {
-  let template = `<svg viewBox='${data.$.viewBox}' xmlns='http://www.w3.org/2000/svg' width='{{svgSize}}px' height='{{svgSize}}px'>`;
-
-  for (const domName of Object.keys(data)) {
-    if (domName === '$') {
-      continue;
-    }
-
-    const counter = {
-      colorIndex: 0,
-    };
-
-    if (data[domName].$) {
-      template += `<${domName}${addAttribute(domName, data[domName], counter)} />`;
-    } else if (Array.isArray(data[domName])) {
-      data[domName].forEach((sub) => {
-        template += `<${domName}${addAttribute(domName, sub, counter)} />`;
-      });
-    }
-  }
-
-  template += `</svg>`;
-
-  return template;
-};
-
-const addAttribute = (domName: string, sub: XmlData['svg']['symbol'][number]['path'][number], counter: { colorIndex: number }) => {
-  let template = '';
-
-  if (sub && sub.$) {
-    if (ATTRIBUTE_FILL_MAP.includes(domName)) {
-      // Set default color same as in iconfont.cn
-      // And create placeholder to inject color by user's behavior
-      sub.$.fill = sub.$.fill || '#333333';
-    }
-
-    for (const attributeName of Object.keys(sub.$)) {
-      if (attributeName === 'fill') {
-        const color = sub.$[attributeName];
-
-        template += ` ${attributeName}='{{(isStr ? color : color[${counter.colorIndex}]) || '${color}'}}'`;
-        counter.colorIndex += 1;
-      } else {
-        template += ` ${attributeName}='${sub.$[attributeName]}'`;
-      }
-    }
-  }
-
-  return template;
-};
