+++ src/libs/generateWechatComponent.ts
@@ -6,14 +6,14 @@ import colors from 'colors';
 import { XmlData } from './fetchXml';
 import { Config } from './getConfig';
 import { getTemplate } from './getTemplate';
+import { generateCase } from "./utils"
 import {
-  replaceHexToRgb, replaceIsRpx,
+   replaceIsRpx,
   replaceNames,
   replaceSize,
 } from './replace';
 // import { whitespace } from './whitespace';
 
-const ATTRIBUTE_FILL_MAP = ['path'];
 
 export const generateWechatComponent = (data: XmlData, config: Config) => {
   const svgTemplates: string[] = [];
@@ -35,7 +35,9 @@ export const generateWechatComponent = (data: XmlData, config: Config) => {
 
     names.push(iconIdAfterTrim);
     svgTemplates.push(
-      `<!--${iconIdAfterTrim}-->\n<view wx:if="{{name === '${iconIdAfterTrim}'}}" style="background-image: url({{quot}}data:image/svg+xml, ${generateCase(item)}{{quot}});` +
+      `<!--${iconIdAfterTrim}-->\n<view wx:if="{{name === '${iconIdAfterTrim}'}}" style="background-image: url({{quot}}data:image/svg+xml, ${generateCase(item, {
+        hexToRgb: true
+      })}{{quot}});` +
       ' width: {{svgSize}}px; height: {{svgSize}}px; " class="icon" />'
     );
 
@@ -60,53 +62,3 @@ export const generateWechatComponent = (data: XmlData, config: Config) => {
   console.log(`\n${colors.green('√')} All icons have been putted into dir: ${colors.green(config.save_dir)}\n`);
 };
 
-const generateCase = (data: XmlData['svg']['symbol'][number]) => {
-  let template = `<svg viewBox='${data.$.viewBox}' xmlns='http://www.w3.org/2000/svg' width='{{svgSize}}px' height='{{svgSize}}px'>`;
-
-  for (const domName of Object.keys(data)) {
-    if (domName === '$') {
-      continue;
-    }
-
-    const counter = {
-      colorIndex: 0,
-    };
-
-    if (data[domName].$) {
-      template += `<${domName}${addAttribute(domName, data[domName], counter)} />`;
-    } else if (Array.isArray(data[domName])) {
-      data[domName].forEach((sub) => {
-        template += `<${domName}${addAttribute(domName, sub, counter)} />`;
-      });
-    }
-  }
-
-  template += `</svg>`;
-
-  return template.replace(/<|>/g, (matched) => encodeURI(matched));
-};
-
-const addAttribute = (domName: string, sub: XmlData['svg']['symbol'][number]['path'][number], counter: { colorIndex: number }) => {
-  let template = '';
-
-  if (sub && sub.$) {
-    if (ATTRIBUTE_FILL_MAP.includes(domName)) {
-      // Set default color same as in iconfont.cn
-      // And create placeholder to inject color by user's behavior
-      sub.$.fill = sub.$.fill || '#333333';
-    }
-
-    for (const attributeName of Object.keys(sub.$)) {
-      if (attributeName === 'fill') {
-        const color = replaceHexToRgb(sub.$[attributeName]);
-
-        template += ` ${attributeName}='{{(isStr ? colors : colors[${counter.colorIndex}]) || '${color}'}}'`;
-        counter.colorIndex += 1;
-      } else {
-        template += ` ${attributeName}='${sub.$[attributeName]}'`;
-      }
-    }
-  }
-
-  return template;
-};
